delivery_note_generator/requirements.txt
pandas>=2.2.0
weasyprint>=60.0
python-barcode>=0.15.0
Pillow>=10.0.0
